requirements.txt
# Core
fastapi[standard]==0.115.0
uvicorn==0.32.0
chainlit==2.10.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0

# Optional: SFT training
# transformers>=4.45.0
# peft>=0.13.0
# trl>=0.11.0
# datasets>=3.0.0
# torch>=2.4.0
